refactor(choose play): introduce temp variable optimal_spells
I'm a little worried this isn't readable enough, but I think it
embodies an important concept:

We order castable spells in descending CMC order. We often want to know
things about "the spell that would use all our mana," or "the spell that
would use all but one of our mana." The optimal_spells list has the
former at index 0 and the latter at index 1.

Maybe I'll think of a more elegant way to handle this later...